Convert GO/Reactome Pathway IDs to Full Names

Converts GO or Reactome IDs to their respective full geneset or pathway names.

RenameGO(term, add_id = T, add_n_gene = T, spe = getOption("spe"))

RenameReactome(term, add_id = T, add_n_gene = T, spe = getOption("spe"))

Arguments

term

GO/Reactome IDs, which can be either a character vector or a matrix (with IDs in rownames).

add_id

Logical: Should the IDs be displayed? Default: TRUE.

add_n_gene

Logical: Should the number of genes in each geneset be displayed? Default: TRUE.

spe

Species, either 'human' or 'mouse'. Default: getOption("spe").

Value

Returns either a character vector or a matrix, depending on the input.

Details

For more examples, see GeneSetAnalysis().

Examples

library(SeuratExtend)

RenameGO(c("GO:0002376","GO:0050896"), spe = "human")

RenameReactome(c("R-HSA-109582","R-HSA-112316"), spe = "human")