(Deprecated) Run Palantir Diffusion Map

This is a deprecated version of the Palantir Diffusion Map function. Please refer to the updated function, Palantir.RunDM, for the latest features and improvements. This older version, in addition to calculating the diffusion map (dm) and multiscale_space (ms), also calculates t-SNE. To use the t-SNE feature, the multicore-tsne package needs to be installed in the conda environment.

RunPalantirDiffusionMap(seu, reduction = "pca", n_components = 20)

Arguments

seu

A Seurat object.

reduction

A character string specifying the dimension reduction to use. Default: 'pca'.

n_components

An integer specifying how many dimensions to use as input. Default: 20.

Value

A Seurat object with added dm, ms, and t-SNE embeddings.

See also

Examples

if (FALSE) {
library(Seurat)
library(SeuratExtend)
pbmc <- RunPalantirDiffusionMap(pbmc)
DimPlot(pbmc, reduction = "tsne", label = TRUE)
}